+++ lib/src/languages/java/java11/java11_parser.dart
@@ -2,9 +2,7 @@
 // This code is governed by the Apache License, Version 2.0.
 // Please refer to the LICENSE and AUTHORS files for details.
 
-import '../../../apollovm_base.dart';
 import '../../../apollovm_parser.dart';
-import '../../../apollovm_runner.dart';
 import 'java11_grammar.dart';
 
 /// Java11 implementation of an [ApolloParser].
@@ -25,17 +23,5 @@ class ApolloParserJava11 extends ApolloParser {
     }
 
     return false;
-  }
-}
-
-class ApolloRunnerJava11 extends ApolloLanguageRunner {
-  ApolloRunnerJava11(ApolloVM apolloVM) : super(apolloVM);
-
-  @override
-  String get language => 'java11';
-
-  @override
-  ApolloRunnerJava11 copy() {
-    return ApolloRunnerJava11(apolloVM);
   }
 }
